Adjusting code to my taste (and C++ ussuals).

+++ src/arff_scanner.h
@@ -54,13 +54,13 @@ public:
      * @brief Returns the current line position
      * @return current line
      */
-    int64 line() const;
+    unsigned long long line() const;
 
     /**
      * @brief Returns the current column position
      * @return current column
      */
-    int64 column() const;
+    unsigned long long column() const;
 
     /**
      * @brief Whether the file has reached end or not
@@ -93,9 +93,9 @@ private:
     /** file being read */
     std::string m_file;
     /** current line being read */
-    int64 m_line;
+    unsigned long long m_line;
     /** current position in the row being read */
-    int64 m_col;
+    unsigned long long m_col;
     /** current character read from the file */
     char m_char;
     /** previous character read from the file */
